+++ SimpleTableManager/SimpleTableManager/Models/IShape1Sized.cs
@@ -2,5 +2,5 @@ namespace SimpleTableManager.Models;
 
 public interface IShape1Sized
 {
-	decimal Size1 { get; set; }
+	double Size1 { get; set; }
 }
