+++ SimpleTableManager/SimpleTableManager/Models/Shape2Sized2dBase.cs
@@ -6,11 +6,11 @@ namespace SimpleTableManager.Models;
 public abstract class Shape2Sized2dBase<T> : Shape1Sized2dBase<T>, IShape2Sized, IParseCore<T>
 where T : Shape2Sized2dBase<T>, IParsable<T>, new()
 {
-	public decimal Size2 { get; set; }
+	public double Size2 { get; set; }
 
 	public Shape2Sized2dBase() { }
 
-	public Shape2Sized2dBase(decimal size1, decimal size2) : base(size1)
+	public Shape2Sized2dBase(double size1, double size2) : base(size1)
 	{
 		Size2 = size2;
 	}
@@ -27,8 +27,8 @@ where T : Shape2Sized2dBase<T>, IParsable<T>, new()
 
 	public new static T ParseCore(GroupCollection args)
 	{
-		var size1 = decimal.Parse(args["s"].Value);
-		var size2 = args["s2"].Success ? decimal.Parse(args["s2"].Value) : size1;
+		var size1 = double.Parse(args["s"].Value);
+		var size2 = args["s2"].Success ? double.Parse(args["s2"].Value) : size1;
 
 		return new T() { Size1 = size1, Size2 = size2 };
 	}
